requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
pillow==11.2.1
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
pystray==0.19.5
python-dotenv==1.1.1
python-xlib==0.33
requests==2.32.4
ruff==0.12.0
shiboken6==6.9.1
six==1.17.0
urllib3==2.5.0
